Fixed thumbnails not showing up when zoom is enabled
Fixed a bug that caused thumbnails to not show up in some cases,
specially when zoom was enabled. It was caused by the variable
in_nextprev not really being set to 0 in some calls to the
backend, even if there were comments saying it would be the case.
Combined with a previous commit, this should also address Debian
bug #379869

Addresses-Debian-Bug: #214112
Signed-off-by: Diego Brea <diego.brea@gmx.com>