Silence `misleading-indentation` warning from gcc
gcc emits a `misleading-indentation` warning over three pairs of `if`s
in `dither.c`, guessing that the second `if` in each pair could be
misconstrued as being guarded by the first.

Adding an `else` silences this warning, and also makes it clearer that
both `if`s work in tandem to clamp each value within a 0-255 range.