Remember our own current path location, before symlink resolution
Navigating across symlinked directories can be somewhat confusing with
xzgv, as `chdir()` will resolve the link when going down such a
directory, so that going back up via the parent may not return to the
original location.

File browsers will typically maintain their own internal path location,
removing the last element when asked to go up to the parent directory.
This is what we're doing here, replacing `chdir()` and `getcwd()` with
wrappers of our own.